refactor: change all directory and create index.jsx files
