--- src/components/ControlPanel/Button/styled.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const StyledButton = styled.button`
-    color: black;
-    padding: 1rem 0.5rem;
-    cursor: pointer;
-`;
-
-export default StyledButton;
