--- src/components/ControlPanel/TogglerDivide/styled.js
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledTogglerDivide = styled.div`
-    display: flex;
-    gap: 2rem;
-    align-items: center;
-    padding: 1rem;
-`;
-
-export const Label = styled.label`
-    display: block;
-    width: 30px;
-    height: 30px;
-    cursor: pointer;
-`;
