+++ src/components/Display/DisplayFC.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import HistoryDisplay from '../HistoryDisplay/HistoryDisplayFC';
-
+import HistoryDisplay from './HistoryDisplay/HistoryDisplayFC';
 import WrapperDisplay from './styled';
 
 const Display = () => {
