+++ src/components/SettingsList/ClearHistory/index.jsx
