+++ src/components/SettingsList/SwitcherTheme/index.jsx
