+++ src/components/SettingsList/config.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import ClearButton from './ClearHistory/ButtonFC';
-import TogglerHistory from './ShowHistory/TogglerHistory';
-import SwitcherThemeFC from './SwitcherTheme/Switcher';
+import ClearButton from './ClearHistory';
+import TogglerHistory from './ShowHistory';
+import SwitcherThemeFC from './SwitcherTheme';
 
 const SETTING_LIST = [
     {
