+++ src/pages/Settings/SettingsFC.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Header from '@components/Header/HeaderFC';
-import SettingsListFC from '@components/SettingsList/SettingsListFC';
+import SettingsList from '@components/SettingsList';
 
 import { WrapperSettings } from './styled';
 
 const SettingsFC = () => (
     <WrapperSettings>
         <Header />
-        <SettingsListFC />
+        <SettingsList />
     </WrapperSettings>
 );
